Refactor : delete unused files and rename files
리팩토링 전 DicomTable 및 관련 컴포넌트들을 삭제하고
리팩토링 후 몇몇 Data Grid 관련 컴포넌트명을 기존 컴포넌트명으로 변경합니다.

Task: #270
