+++ client/src/Page.js
@@ -2,7 +2,6 @@ import * as React from 'react';
 import Box from '@mui/material/Box';
 import CssBaseline from '@mui/material/CssBaseline';
 import { ButtonGroup, CircularProgress,Stack,Typography,Button } from '@mui/material';
-import DicomTable from "./component/Table/DicomTable";
 import UpDownloadToolbar from "./component/Toolbar/UpDownloadToolbar";
 import { DrawerHeader } from './component/Drawer/ProjectDrawer';
 import ProjectDrawer from './component/Drawer/ProjectDrawer'
@@ -168,19 +167,10 @@ export default function Page() {
                                     </Typography>
                                 </Stack>
                                 :
-                                <Stack>
-                                    <MetaDataGrid
-                                        metaData={metaData}
-                                        project={presentProject}
-                                    />
-                                    <DicomTable 
-                                        data={metaData} 
-                                        metaDataUpdated={metaDataUpdated}
-                                        setMetaDataUpdated={setMetaDataUpdated}
-                                        isNonReferenced={presentProject.projectName === 'Non-Reference Dicom' ? true : false}
-                                        project={presentProject}
-                                    />
-                                </Stack>
+                                <MetaDataGrid
+                                    metaData={metaData}
+                                    project={presentProject}
+                                />
                             }
                         </div>
                         :<div></div>
